Move getSmartModuleName inline with legacy DefaultParser
