+++ packages/hekla-core/src/legacy/parsers/DefaultParser/index.js
@@ -1,6 +1,8 @@
 'use strict';
 
 const path = require('path');
+const camelCase = require('camel-case');
+
 const astUtils = require('../../../utils/ast-utils');
 const fsUtils = require('../../../utils/fs-utils');
 const BaseParser = require('../BaseParser');
@@ -29,7 +31,7 @@ function analyzeDefaultComponent(ast, module) {
 
 function getComponentDetails(module, ast) {
   return {
-    name: fsUtils.getSmartModuleName(module.path),
+    name: getSmartModuleName(module.path),
     type: 'default',
     path: module.path,
     dependencies: getDependencies(ast)
@@ -39,4 +41,29 @@ function getComponentDetails(module, ast) {
 function getDependencies(ast) {
   return astUtils.getImportInfo(ast)
     .map(importItem => importItem.value);
+}
+
+function getSmartModuleName(filePath) {
+  const pathPieces = filePath.split('/');
+  const filename = pathPieces[pathPieces.length - 1];
+  const directory = _maybeCamelCase(pathPieces[pathPieces.length - 2]);
+
+  const filePieces = filename.split('.');
+  const filenameNoExt = _maybeCamelCase(filePieces[0]);
+
+  if (filenameNoExt === 'index') {
+    return directory;
+  } else if (filenameNoExt === 'app') {
+    return directory + 'App';
+  } else {
+    return filenameNoExt;
+  }
+}
+
+function _maybeCamelCase(name) {
+  if (name.indexOf('-')) {
+    return camelCase(name);
+  } else {
+    return name;
+  }
 }
