+++ packages/hekla-core/src/utils/fs-utils/index.spec.js
@@ -1,6 +1,3 @@
-'use strict';
-
-const path = require('path');
 const fsUtils = require('./index');
 
 describe('fsUtils', () => {
@@ -46,35 +43,6 @@ describe('fsUtils', () => {
     it('should shorten a module named app.js', () => {
       const moduleName = './src/my-feature/app.js';
       expect(fsUtils.getModuleShortName(moduleName)).to.equal('my-feature/app.js');
-    });
-
-  });
-
-  describe('getSmartModuleName', () => {
-
-    it('should handle a normal module name', () => {
-      const filePath = '/path/to/normalModule.js';
-      expect(fsUtils.getSmartModuleName(filePath)).to.equal('normalModule');
-    });
-
-    it('should handle a module name with extra file extensions', () => {
-      const filePath = '/path/to/normalModule.thing.js';
-      expect(fsUtils.getSmartModuleName(filePath)).to.equal('normalModule');
-    });
-
-    it('should handle a module name with hyphens', () => {
-      const filePath = '/path/to/normal-module.js';
-      expect(fsUtils.getSmartModuleName(filePath)).to.equal('normalModule');
-    });
-
-    it('should handle a module named index.js', () => {
-      const filePath = '/path/to/normalModule/index.js';
-      expect(fsUtils.getSmartModuleName(filePath)).to.equal('normalModule');
-    });
-
-    it('should handle a module named app.js', () => {
-      const filePath = '/path/to/normalModule/app.js';
-      expect(fsUtils.getSmartModuleName(filePath)).to.equal('normalModuleApp');
     });
 
   });
